flask_rsa_encryption/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Naked==0.1.32
pycryptodome==3.19.0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
urllib3==2.0.6
Werkzeug==3.0.0